git-happens/pyproject.toml
[project]
name = "git-happens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Simon Simonsson", email = "simon.simonsson@posten.no" }
]
requires-python = ">=3.11"
dependencies = [
    "databricks-sdk[openai]>=0.57.0",
    "mcp>=1.9.4",
    "openai>=1.91.0",
    "smolagents[mcp]>=1.19.0",
]

[project.scripts]
git-happens = "git_happens:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipython>=9.3.0",
]
